+++ server/src/estimating/mod.rs
@@ -5,6 +5,5 @@ pub mod estimate_assembly;
 pub mod project;
 
 pub use assembly::Assembly;
-pub use estimate::Estimate;
 pub use estimate_assembly::EstimateAssembly;
 pub use project::Project;
