+++ server/src/http/estimate/loader.rs
@@ -1,4 +1,4 @@
-use crate::estimating::Estimate;
+use crate::http::estimate::Estimate;
 use async_graphql::dataloader::Loader;
 use async_graphql::FieldError;
 use async_trait::async_trait;
