+++ server/src/http/loaders/mod.rs
@@ -1,9 +1,8 @@
 mod assembly_items;
 mod estimate_assemblies;
-mod estimates;
 mod projects;
 
+pub use crate::http::estimate::loader::EstimateLoader;
 pub use assembly_items::AssemblyItemLoader;
 pub use estimate_assemblies::EstimateAssembliesLoader;
-pub use estimates::EstimateLoader;
 pub use projects::ProjectLoader;
